CASM-3706: Skip uploading already-existing images
This change modifies the behavior of the
`ImsHelper.image_upload_artifacts()` method to skip uploading images
which have already been uploaded. The method now checks if there is an
image in IMS which has the same name as the desired image, and if it has
associated artifacts, then it is skipped. Empty images are uploaded and
patched.

Test Description:
Test on internal system by uploading artifacts to empty images, creating
new images, and attempting to re-upload existing images using a test
harness script.